requirements.txt
amqp==5.3.1
asgiref==3.9.2
billiard==4.2.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
docopt==0.6.2
idna==3.10
kombu==5.5.4
packaging==25.0
pipreqs==0.4.13
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
yarg==0.1.10
